Restructured core functionality
